refactor: make refactor in transfer component
